requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data processing and augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Web application
streamlit>=1.28.0
streamlit-option-menu>=0.3.0

# Experiment tracking
clearml>=1.13.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
tqdm>=4.64.0
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0